requirements.txt
Django==5.1.3
djangorestframework==3.15.2
pillow==11.0.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
requests